Rename dir and components
